Attach additional properties error to each bad key
Instead of a single error for unexpected additional properties attached
to the container dict()'s path, create an error for each unexpected key
and include the key in the path. This is useful when parsing errors to
identify the exact failing object, which is helpful when identifying
the location of the failing key.

In the particular case of using the ruamel.yaml parser, line & column
numbers are preserved in the data structure, and having the full path to
the failing key means error messages can include the line & column
numbers.

Signed-off-by: Grant Likely <grant.likely@arm.com>